+++ Token.js
@@ -32,14 +32,19 @@ function isConstrainedTokenBorder() { return !ConstrainedTokenBorder.get(this)._
  * Determine the correct border shape for this token. Utilize the cached token shape.
  * @returns {PIXI.Polygon|PIXI.Rectangle}
  */
-function tokenBorder() { return this.tokenShape.translate(this.document.x, this.document.y); }
+function tokenBorder() { return this.shape.translate(this.document.x, this.document.y); }
 
 /**
  * New getter: Token.prototype.tokenShape
  * Cache the token shape.
  * @type {PIXI.Polygon|PIXI.Rectangle}
  */
-function tokenShape() { return this._tokenShape || (this._tokenShape = calculateTokenShape(this)); }
+function tokenShape() {
+  const msg = "libGeometry|Token#tokenShape is deprecated. "
+    + "If you need the shape of a Token, use Token#shape/getShape instead.";
+  foundry.utils.logCompatibilityWarning(msg, {since: 12, until: 14, once: true});
+  return this.shape;
+}
 
 PATCHES.CONSTRAINED_TOKEN_BORDER.GETTERS = {
   constrainedTokenBorder,
@@ -47,22 +52,3 @@ PATCHES.CONSTRAINED_TOKEN_BORDER.GETTERS = {
   tokenShape,
   isConstrainedTokenBorder
 };
-
-// ----- NOTE: Helper functions ----- //
-/**
- * Theoretical token shape at 0,0 origin.
- * @returns {PIXI.Polygon|PIXI.Rectangle}
- */
-function calculateTokenShape(token) {
-  // TODO: Use RegularPolygon shapes for use with WeilerAtherton
-  // Hexagon (for width .5 or 1)
-  // Square (for width === height)
-  let shape;
-  if ( canvas.grid.isHexagonal ) {
-    const pts = canvas.grid.grid.getBorderPolygon(token.document.width, token.document.height, 0);
-    if ( pts ) shape = new PIXI.Polygon(pts);
-  }
-
-  return shape || new PIXI.Rectangle(0, 0, token.w, token.h);
-}
-